Complete authorization

+++ controlers/goalControler.js
@@ -1,10 +1,10 @@
 const asyncHandler = require('express-async-handler')
-
 const Goal = require('../model/goalModel')
+const User = require('../model/userModel')
 
 // Get goals // GET /api/goals // Private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id})
     res.json(goals)
 })
 
@@ -16,7 +16,8 @@ const setGoal = asyncHandler(async (req, res) => {
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
 
     res.json(goal)
@@ -30,6 +31,20 @@ const updateGoal = asyncHandler(async (req, res) => {
         throw new Error('Goal not found')
     } 
 
+    const user = await User.findById(req.user.id)
+
+    //Check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    //Make sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true})
     res.json(updatedGoal)
 })
@@ -41,6 +56,20 @@ const deleteGoal = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error('Goal not found')
     } 
+
+    const user = await User.findById(req.user.id)
+
+    //Check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    //Make sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
+    }
 
     await goal.deleteOne()
